+++ lib/ui/src/lib.rs
@@ -3,4 +3,6 @@ pub mod draw;
 pub mod frame;
 pub mod login_window;
 pub mod rect;
+pub mod server_list_window;
+pub mod state;
 pub mod text_input;